requirements.txt
face_recognition==1.3.0
opencv-python==4.7.0.72
numpy==1.23.5
pillow==9.4.0
pyyaml==6.0
pandas==1.5.3
